+++ src/Mock/SluggerInterfaceMock.php
@@ -13,6 +13,7 @@ use Mockery\MockInterface;
 use PrecisionSoft\Symfony\Phpunit\Container\MockContainer;
 use PrecisionSoft\Symfony\Phpunit\Contract\MockDtoInterface;
 use PrecisionSoft\Symfony\Phpunit\MockDto;
+use Symfony\Component\String\AbstractUnicodeString;
 use Symfony\Component\String\Slugger\SluggerInterface;
 use Symfony\Component\String\UnicodeString;
 
@@ -33,7 +34,7 @@ class SluggerInterfaceMock implements MockDtoInterface
         return static function (MockInterface $mockInterface, MockContainer $mockContainer): void {
             $mockInterface->shouldReceive('slug')
                 ->byDefault()
-                ->andReturnUsing(static function (string $string): UnicodeString {
+                ->andReturnUsing(static function (string $string): AbstractUnicodeString {
                     return new UnicodeString($string);
                 });
         };