+++ tests/bootstrap.php
@@ -7,3 +7,9 @@ declare(strict_types=1);
  */
 
 require \dirname(__DIR__) . '/vendor/autoload.php';
+
+\register_shutdown_function(static function (): void {
+    if (true === \class_exists(\PrecisionSoft\Symfony\Phpunit\Test\Utility\TestKernel::class)) {
+        \PrecisionSoft\Symfony\Phpunit\Test\Utility\TestKernel::cleanupTempDirs();
+    }
+});